pyproject.toml
[project]
name = "NoBrainerRag"
version = "0.1.2"
description = "A dead simple RAG pipeline with persistent memory and conversation namespaces."
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Aarush Srivatsa", email = "your.email@example.com"}
]
keywords = ["rag", "retrieval", "llm", "embeddings", "pinecone", "langchain", "vector-database"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "langchain>=0.1.0",
    "langchain-community>=0.0.10",
    "langchain-text-splitters>=0.0.1",
    "langchain-pinecone>=0.0.1",
    "flashrank>=0.2.0",
    "pinecone-client>=3.0.0",
    "python-dotenv>=1.0.0",
    "langchain-ollama>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/AarushSrivatsa/NoBrainerRag"
Repository = "https://github.com/AarushSrivatsa/NoBrainerRag"
Issues = "https://github.com/AarushSrivatsa/NoBrainerRag/issues"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["NoBrainerRag"]